Model Controller complete

+++ AirQuality ObjC/AirQuality ObjC/Models/DVMWeather.h
@@ -14,9 +14,9 @@ NS_ASSUME_NONNULL_BEGIN
 
 @property (nonatomic, readonly) NSInteger temperature;
 @property (nonatomic, readonly) NSInteger humidity;
-@property (nonatomic, readonly) NSInteger windSpeed;
+@property (nonatomic, readonly) float windSpeed;
 
--(instancetype)initWithWeatherInfo:(NSInteger)temperature humidity:(NSInteger)humidity windSpeed:(NSInteger)windSpeed;
+-(instancetype)initWithWeatherInfo:(NSInteger)temperature humidity:(NSInteger)humidity windSpeed:(float)windSpeed;
 
 @end
